funcionalidad 2 lista
